refactor(test): merge compose root endpoint tests into route.test.ts (#9393)
Consolidate 4 non-route test files (get-by-name, instructions-volume,
sandbox-capability, upsert) into a single route.test.ts for the
/api/agent/composes endpoint, aligning with the Web Route Integration
Test architecture. All 40 test cases preserved with no behavior changes.
